+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # uwsgiconf changelog
 
+### Unreleased
+* ** Django.contrib. Fixed task forced task run missing tasks message (fixes #16).
+
 ### v2.2.1 [2025-09-27]
 * ** Django. Admin. Task representation improved.